pyproject.toml
[project]
name = "unitconvertorproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.10",
    "streamlit>=1.43.1",
]
